Crate problem108

Problem 108 (Diophantine reciprocals I)

Solution detail

Firstly, some rearrangement of the equation is in order. The given equation can equivalently be written in the form (x - n)(y - n) = n².

Given that interchanging x, y does not count as a new solution, the number of solutions to this equation, and hence to the original equation, is given by (d(n²) + 1) / 2, where d(n²) is the number of divisors of n².

We are therefore searching for the smallest n for which d(n²) is at least 2001 (in general, greater than 2 * target where target is the number of solutions we want to be in excess of).

This can be done efficiently using a depth-first search through numbers having the following properties:

They are smooth, in the sense that they are divisible exactly by primes 2, 3, ..., p for some prime p.

They are descending, in the sense that the exponents of their prime factors are strictly decreasing as the prime factor increases.

Given these two properties, a number is uniquely determined by the sequence of exponents in its prime factorisation, so we can conduct the search on those without worrying about actual primes, and only convert to the actual solution at the end.

Given a sequence of exponents, the number of divisors of the square of that number is equal to the product of 2e + 1 over all exponents e.

It is therefore easy to conduct a depth-first search over all decreasing sequences of exponents, taking those nodes which correspond to a number whose square has enough divisors, and finding, from those nodes, the one which corresponds to the smallest actual number.

To spped the search up, the smallest solution so far can be stored as auxiliary data and nodes pruned whenever the value of the node exceeds the smallest solution found so far.

Structs

SquareDivisorsTree
The state of the search for numbers whose squares have sufficiently many divisors.

SquareDivisorsTreeStep
A step that can be taken in the search tree

Functions

answer
Solve the problem, returning the answer as a String

main

solve
Find the smallest n for which the equation has more than target solutions.